2011-02-22  Andreas Kling  <kling@webkit.org>
        Reviewed by Dan Bernstein.

        FontCache: Make cTargetInactiveFontData an int instead of float.
        https://bugs.webkit.org/show_bug.cgi?id=54963

        * platform/graphics/FontCache.cpp:

Canonical link: https://commits.webkit.org/69322@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@79402 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/FontCache.cpp
@@ -261,7 +261,7 @@ typedef HashMap<FontPlatformData, pair<SimpleFontData*, unsigned>, FontDataCache
 static FontDataCache* gFontDataCache = 0;
 
 const int cMaxInactiveFontData = 120;  // Pretty Low Threshold
-const float cTargetInactiveFontData = 100;
+const int cTargetInactiveFontData = 100;
 static ListHashSet<const SimpleFontData*>* gInactiveFontData = 0;
 
 SimpleFontData* FontCache::getCachedFontData(const FontDescription& fontDescription, const AtomicString& family, bool checkingAlternateName)
